Conflicts:
	www/modules/files/CustomField.js
	www/modules/files/Portlet.js
	www/modules/files/PortletExpiredFiles.js
	www/modules/files/controller/FolderController.php
	www/modules/files/model/File.php
	www/modules/files/model/Folder.php
	www/views/Extjs3/javascript/go/util.js
	www/views/Extjs3/javascript/scripts.txt
	www/views/Extjs3/themes/Paper/src/_go-style.scss

+++ www/go/core/util/Lock.php
@@ -75,7 +75,9 @@ class Lock {
 		//cleanup lock file if lock() was used
 		if(isset($this->lockFile)) {
 			fclose($this->lockFp);
-			unlink($this->lockFile);			
+			if(file_exists($this->lockFile)) {
+				unlink($this->lockFile);			
+			}
 		}
 	}
